requirements.txt
langgraph>=0.2.36
pdfplumber>=0.11.4
python-docx>=1.1.2
sentence-transformers>=3.0.1
transformers>=4.41.0
faiss-cpu>=1.8.0
numpy>=1.26.0
networkx>=3.3
matplotlib>=3.8.0
